Support Android Gradle Plugin 3.0.0+ (#166)
Instead of grabbing variant output from the assemble task, we can hook up to the package task and depend on that.

The wrinkle is that, in the new tools version, its output is a directory, as opposed to the prior version that yielded a single package file. The counter task is a little more complicated now, but things seem to work. Fingers crossed.

Fixes #161